Restructure folder to show Vue as main App
