+++ cypress/support/step_definitions/navigate.js
@@ -1,6 +1,6 @@
 import { Given } from "cypress-cucumber-preprocessor/steps";
 
-const url = "http://127.0.0.1:8080/";
+const url = "http://localhost:8080/";
 Given("I open the website", () => {
   cy.visit(url);
 });
