+++ package.json
@@ -1,38 +1,57 @@
 {
-  "name": "eddie",
-  "version": "1.0.0",
+  "name": "vuejs_version",
+  "version": "0.1.0",
+  "private": true,
   "description": "## What is our community about - We believe open source is for everyone. - Our community is inclusive and a safe space for all. - Please welcome new people. - The best way to learn is to help someone else. - [Code Of Conduct](https://github.com/EddieJaoudeCommunity/EddieJaoudeCommunity.github.io/blob/main/CODE_OF_CONDUCT.md)",
-  "main": "index.js",
   "cypress-cucumber-preprocessor": {
     "nonGlobalStepDefinitions": false
   },
   "scripts": {
     "test": "cypress run",
     "cypress:open": "cypress open",
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --port 8080"
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/EddieJaoudeCommunity/EddieJaoudeCommunity.github.io.git"
   },
-  "keywords": [],
   "author": "Eddie Jaoude Community",
   "license": "MIT",
   "homepage": "https://github.com/EddieJaoudeCommunity/EddieJaoudeCommunity.github.io#readme",
   "devDependencies": {
     "cypress": "^5.2.0",
     "cypress-cucumber-preprocessor": "^2.5.5",
-    "@rollup/plugin-commonjs": "^14.0.0",
-    "@rollup/plugin-node-resolve": "^8.0.0",
-    "rollup": "^2.3.4",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^6.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.0.0"
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0-0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.7.2",
+    "eslint-plugin-vue": "^7.0.0-0"
   },
   "dependencies": {
-    "sirv-cli": "^1.0.0"
-  }
+    "core-js": "^3.6.5",
+    "vue": "^3.0.0-0"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "babel-eslint"
+    },
+    "rules": {}
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead"
+  ]
 }
