+++ src/main.js
@@ -1,7 +1,4 @@
-import App from './App.svelte';
+import { createApp } from 'vue'
+import App from './App.vue'
 
-const app = new App({
-	target: document.body
-});
-
-export default app;
+createApp(App).mount('#app')
